Configure coverage helpers for test exec group
They are run on the test action's execution platform, which is resolved for the `test` exec group, not the default one.

Since this adds a reference to the `test` exec group, which is only defined for test rules, `py_executable` attrs must not merge in coverage attributes if the rule to be created is not a test rule.