wip: reimplement namespace pkg filepath calc in starlark
With this PR I would like to facilitate the implementation of the venv
layouts because we can in theory take the `srcs` and the `data` within
the `py_library` and then use the `expand_template` to write the extra
Python files if the namespace_pkgs flag is enabled.

This is WIP so that maintainers can hop in if they have time.

TODO:
- [ ] Do the ctx.actions.expand_template based on the output from the
  reimplementation
- [ ] Thread the namespace_pkgs attr
- [ ] Remove Python code? Put under pipstar?

Work towards #2156