docs/refactor: Use python.defaults, not is_default
When there are multiple Python toolchains, there are currently two
ways of setting the default version: the is_default attribute of
python.toolchain() tag class and the python.defaults() tag class. The
latter is more powerful, since it also supports files and environment
variables. This patch updates the examples, docs and the MODULE.bazel
file to use python.defaults().

Relates to pull request #2588 and issue #2587.