feat: support arbitrary config_setting labels in our platforms 3/n
With this PR we can support arbitrary target settings instead of just
plain `constraint_values`. We still have custom logic to ensure that all
of the tests pass. However, the plan is to remove those tests once we
have simplified the wheel selection mechanisms and the `pkg_aliases`
macro. I.e. if we have at most 1 wheel per platform that the `pypi`
bzlmod extension passes to the `pkg_aliases` macro, then we can just
have a simple `selects.with_or` where we list out all of the target
platform values.

This PR may result in us creating more targets but that is the price
that we have to pay if we want to do this incrementally.

The last remaining thing for #2548 is to make sure that we are not
parsing the user-friendly strings to get the `os` and the `cpu`.

Work towards #2747
Work towards #2548
Work towards #260