fix(pypi): pull fewer wheels
Before this we would pull all of the wheels that the user target
configuration would be compatible with and that meant that it was not
customizable. This also meant that there were a lot of footguns in the
configuration where the select statements were not really foolproof.

With this PR we select only those sources that need to be for the
declared configurations.

Freethreaded support should be done by defining extre freethreaded
platforms using the new builder API.

Work towards #2747
Work towards #2759
Work towards #2849