chore: drop old Python interpreters and update tooling
We had references to old versions of the Python interpreter, which most
certainly had security vulnerabilities and bugs and probably were not
used much.

If the users wish, they can include old versions by following our docs,
so this PR is about removing the old versions to make our CI faster. The
only thing that I am concerned is that we may be loosing our ability to
test what happens if we have multiple versions of the `3.X` toolchain
registered to ensure that we are resolving to the right interpreter.

Summary:
- [x] Remove old versions.
- [x] Update the tooling to format the printed text to be the exact
  replica of the contents of the `TOOL_VERSIONS` which makes it easier
  to do bulk updates - just copy the output from the terminal and
  cleanup unnecessary lines.
- [x] Document in the Changelog.
- [x] Use the latest python-build-standalone toolchain release.

All of this work is to ensure that it is easier to test and add
pre-releases to catch regressions in the Python behaviour itself.