fix: bootstrapping script to not use multiline f-strings
With bootstrap_impl=system_python, after
https://github.com/bazel-contrib/rules_python/pull/2607,
host Python 3.6 and above is required because of the use of multiline
f-strings. For maximum backwards compatibility (because this is a host
dependency), do not use multiline f-strings.

This issue naturally does not apply to bootstrap_impl=script because
there are no host Python dependency.

Link: https://github.com/bazel-contrib/rules_python/pull/2607