fix(sphinxdocs): add retry logic when exit code 2 occurs
Running Sphinx multiple times in the same process sometimes results
in an error ("exit code 2"). Digging in, this is likely a bug in
the sphinx_bzl plugin in how it merges data when parallel or incremental
builds are performed.

Until that's fixed, work around the problem by internally retrying
the Sphinx build when exit code 2 occurs. This is basically what we're
doing today and should reduce the number of flakes for the RTD builds.