fix: don't call Args.add() with an integer
The documentation for Bazel's Args states that standard conversion rules
are only specified for strings, Files, and Labels. For all other types
the conversion to a string is done in an unspecified manner, which is
why it should be avoided.

Let's stay away from this unspecified behaviour by explicitly converting
the precompile optimization level to a string before calling Args.add().