refactor(runtime setup): Add separate ABI3 Python headers target
Until now, we silently link extensions with both stable  and unstable ABI libs, with the latter
taking precedence in symbol resolution, because it appears first in the linker command AND, crucially,
contains all CPython symbols present in the stable ABI library, thus overriding them.

This has the effect that stable ABI extensions on Windows are useable only with the Python distribution
that they were built on. The now introduced separate ABI3 header target fixes this, and should be
used for C++ extensions on Windows if stable ABI builds are requested.

Idea as formulated by `@dgrunwald-qt` in
https://github.com/nicholasjng/nanobind-bazel/issues/72#issuecomment-3249959583.