fix(pip): do not use experimental_index_url for publish_deps
This usage means that all users will query `PyPI` to get SimpleAPI
metadata. Once bazelbuild/bazel#24777 is merged, we can roll it back,
but then we may need to check-in the `MODULE.bazel.lock` file into git.

This should be no-op for most users because the publisher is usually run
from with `host == exec`, but there is a small chance that people will
hit #2241.

Work towards #2937