fix(venv): group venv prefixes by path component, not raw path
When files overlap between packages, and dist-info directories are
present, it results in a prefix list like `foo foo-bar foo/bar`.
When sorted as raw strings, hyphen sorts before slash, so the continuity
of path prefixes is violated and they are grouped separately. An error
then occurs because both `foo/` and `foo/bar` are created, but the
latter is a sub-path of the former.

To fix, change the sort key to a tuple of path components. This makes
`foo foo-bar foo/bar` sort as `(foo,) (foo, bar), (foo-bar, )`,
resulting in the correct order.

Fixes https://github.com/bazel-contrib/rules_python/issues/3204